Tager det tilbage, get() i DAO returnerer interfaces.
Rettede int til double i amount/mængde.

+++ src/main/java/dal/DTO/Opskrift.java
@@ -78,7 +78,7 @@ public class Opskrift implements IOpskrift {
         return id;
     }
 
-    public ArrayList<Integer> getMaengde() {
+    public ArrayList<Double> getMaengde() {
         return maengde;
     }
